+++ app/Http/Controllers/PagibigContributionController.php
@@ -40,11 +40,11 @@ class PagibigContributionController extends Controller
         $pagibig->fill($request->validated());
         $data = json_decode('{}'); 
         if(!$pagibig->save()){
-            $data->message = "failed to store data";
+            $data->message = "Save failed.";
             $data->success = false;
             return response()->json($data, 400);
         }
-        $data->message = "successfully store data";
+        $data->message = "Successfully save.";
         $data->success = true;
         $data->data = $pagibig;
         return response()->json($data);
@@ -57,15 +57,16 @@ class PagibigContributionController extends Controller
     {
         //
         $pagibig = PagibigContribution::find($id);
-        $data = json_decode('{}'); 
-        $data->message = "successfully show data";
-        $data->success = true;
-        $data->data = $pagibig;
-        if($data->data==null){
-            $data->message = "no data found";
-            $data->success = false;
-        }
-        return response()->json($data);
+        $data = json_decode('{}');
+        if (!is_null($pagibig) ) {
+            $data->message = "Successfully fetch.";
+            $data->success = true;
+            $data->data = $pagibig;
+            return response()->json($data);
+        } 
+        $data->message = "No data found.";
+        $data->success = false;
+        return response()->json($data, 404);
     }
 
     /**
@@ -83,17 +84,22 @@ class PagibigContributionController extends Controller
     {
         //
         $pagibig = PagibigContribution::find($id);
-        $pagibig->fill($request->validated());
         $data = json_decode('{}'); 
-        if($pagibig->save()){
-            $data->message = "successfully update data";
-            $data->success = true;
-            $data->data = $pagibig;
-            return response()->json($data);
+        if (!is_null($pagibig) ) {
+            $pagibig->fill($request->validated());
+            if($pagibig->save()){
+                $data->message = "Successfully update.";
+                $data->success = true;
+                $data->data = $pagibig;
+                return response()->json($data);
+            }
+            $data->message = "Failed update.";
+            $data->success = false;
+            return response()->json($data, 400);
         }
-        $data->message = "failed update data";
+        $data->message = "Failed update.";
         $data->success = false;
-        return response()->json($data, 400);
+        return response()->json($data, 404);
     }
 
     /**
@@ -104,14 +110,19 @@ class PagibigContributionController extends Controller
         //
         $pagibig = PagibigContribution::find($id);
         $data = json_decode('{}'); 
-        if($pagibig->delete()){
-            $data->message = "successfully update data";
-            $data->success = true;
-            $data->data = $pagibig;
-            return response()->json($data);
+        if (!is_null($pagibig) ) {
+            if($pagibig->delete()){
+                $data->message = "Successfully delete.";
+                $data->success = true;
+                $data->data = $pagibig;
+                return response()->json($data);
+            }
+            $data->message = "Failed delete.";
+            $data->success = false;
+            return response()->json($data,400); 
         }
-        $data->message = "failed delete data";
+        $data->message = "Failed delete.";
         $data->success = false;
-        return response()->json($data,400); 
+        return response()->json($data,404); 
     }
 }
